Basic acceptance tested version
This is now ready to sit at the backend, while I play with some frontend
stuff.  Cool stuff (for instance, adding albums as a first class object,
and some form of database for storing metadata/EXIF stuff) can come
later.

The unit tests should be a little more stringent, but we can get to that
later.